### 2023-07-03 #5615 @editorialbot
[PRE REVIEW]: parafields: A generator for distributed, stationary Gaussian processes
**Submitting author:** @dokempf (<a href="http://orcid.org/0000-0002-6140-2332">Dominic Kempf</a>)
**Repository:** https://github.com/parafields/parafields
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.3.0
