### 2023-07-03 #5615 @dokempf
@danielskatz Thanks! Please bear in mind, that the C++ core library https://github.com/parafields/parafields-core/ is also part of this - which increases the code size drastically. In order to keep the legacy C++ use case simple, we decided to keep the C++ and Python code in two separate projects.
