### 2023-07-03 #5615 @danielskatz
ok, then the scope review isn't needed.
