### 2023-07-04 0951193
Correct DOIs for all references
