+++ paper.bib
@@ -5,7 +5,7 @@ volume = {102},
 pages = {161-177},
 year = {2017},
 issn = {0309-1708},
-doi = {https://doi.org/10.1016/j.advwatres.2016.12.006},
+doi = {10.1016/j.advwatres.2016.12.006},
 url = {https://www.sciencedirect.com/science/article/pii/S0309170816307643},
 author = {Ole Klein and Olaf A. Cirpka and Peter Bastian and Olaf Ippisch},
 keywords = {Inverse modeling, Geostatistical inversion, Uncertainty quantification, Nonlinear conjugate gradients, Preconditioning},
@@ -19,7 +19,8 @@ abstract = {In the geostatistical inverse problem of subsurface hydrology, conti
   number={4},
   pages={1088--1107},
   year={1997},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/s1064827592240555}
 }
 @article{graham2018analysis,
   title={Analysis of circulant embedding methods for sampling stationary random fields},
@@ -29,7 +30,8 @@ abstract = {In the geostatistical inverse problem of subsurface hydrology, conti
   number={3},
   pages={1871--1895},
   year={2018},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/17m1149730}
 }
 @misc{parafields-core,
   author={Klein, Ole and Kempf, Dominic},
@@ -78,7 +80,7 @@ pages = {75-112},
 year = {2021},
 note = {Development and Application of Open-source Software for Problems with Numerical PDEs},
 issn = {0898-1221},
-doi = {https://doi.org/10.1016/j.camwa.2020.06.007},
+doi = {10.1016/j.camwa.2020.06.007},
 url = {https://www.sciencedirect.com/science/article/pii/S089812212030256X},
 author = {Peter Bastian and Markus Blatt and Andreas Dedner and Nils-Arne Dreier and Christian Engwer and René Fritze and Carsten Gräser and Christoph Grüninger and Dominic Kempf and Robert Klöfkorn and Mario Ohlberger and Oliver Sander},
 abstract = {This paper presents the basic concepts and the module structure of the Distributed and Unified Numerics Environment and reflects on recent developments and general changes that happened since the release of the first Dune version in 2007 and the main papers describing that state Bastian etal. (2008a, 2008b). This discussion is accompanied with a description of various advanced features, such as coupling of domains and cut cells, grid modifications such as adaptation and moving domains, high order discretizations and node level performance, non-smooth multigrid methods, and multiscale methods. A brief discussion on current and future development directions of the framework concludes the paper.}
@@ -128,5 +130,6 @@ abstract = {This paper presents the basic concepts and the module structure of t
   number={4},
   pages={409--432},
   year={1994},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.2307/1390903}
 }
