### 2023-07-18 #5615 @max-little
Yes sure, but it will take me about 3 weeks.

On Tue, 18 Jul 2023 at 22:37, Patrick Diehl ***@***.***>
wrote:

> Hi @max-little <https://github.com/max-little> do you have time to review
> this paper?
>
