### 2023-07-18 #5615 @diehlpk
> Yes sure, but it will take me about 3 weeks.
> […](#)
> On Tue, 18 Jul 2023 at 22:37, Patrick Diehl ***@***.***> wrote: Hi @max-little <https://github.com/max-little> do you have time to review this paper? — Reply to this email directly, view it on GitHub <[#5615 (comment)](https://github.com/openjournals/joss-reviews/issues/5615#issuecomment-1641028114)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AYOOCSAYH52PTA7P4DOPEXLXQ36YLANCNFSM6AAAAAAZ4TBHU4> . You are receiving this because you were mentioned.Message ID: ***@***.***>
> -- Max Little ([www.maxlittle.net](http://www.maxlittle.net)) Associate Professor, University of Birmingham Turing Fellow, Alan Turing Institute TED Fellow (fellows.ted.com/profiles/max-little) Visiting Associate Professor, MIT Author: Machine Learning for Signal Processing, Oxford University Press global.oup.com/academic/product/machine-learning-for-signal-processing-9780198714934 Room 138, School of Computer Science University of Birmingham Birmingham B15 2TT UK +44 7710 609564 Skype: dr.max.little

That would be fine.
