### 2023-08-04 #5735 @editorialbot
[REVIEW]: parafields: A generator for distributed, stationary Gaussian processes
**Submitting author:** @dokempf (<a href="http://orcid.org/0000-0002-6140-2332">Dominic Kempf</a>)
**Repository:** https://github.com/parafields/parafields
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.1
