### 2023-09-01 fade36c
Bump pypa/cibuildwheel from 2.14 to 2.15
Bumps [pypa/cibuildwheel](https://github.com/pypa/cibuildwheel) from 2.14 to 2.15.
- [Release notes](https://github.com/pypa/cibuildwheel/releases)
- [Changelog](https://github.com/pypa/cibuildwheel/blob/main/docs/changelog.md)
- [Commits](https://github.com/pypa/cibuildwheel/compare/v2.14...v2.15)

---
updated-dependencies:
- dependency-name: pypa/cibuildwheel
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
