+++ .github/workflows/pypi.yml
@@ -24,7 +24,7 @@ jobs:
     - uses: actions/checkout@v3
 
     - name: Build wheels (arch=${{ matrix.arch }})
-      uses: pypa/cibuildwheel@v2.14
+      uses: pypa/cibuildwheel@v2.15
       env:
         CIBW_ARCHS: "${{ matrix.arch }}"
 
