### 2023-10-01 a39e111
Bump pypa/cibuildwheel from 2.15 to 2.16
Bumps [pypa/cibuildwheel](https://github.com/pypa/cibuildwheel) from 2.15 to 2.16.
- [Release notes](https://github.com/pypa/cibuildwheel/releases)
- [Changelog](https://github.com/pypa/cibuildwheel/blob/main/docs/changelog.md)
- [Commits](https://github.com/pypa/cibuildwheel/compare/v2.15...v2.16)

---
updated-dependencies:
- dependency-name: pypa/cibuildwheel
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
