### 2023-10-04 ee7b1c4
Fix incompatibility of nbstripout and pretty-format-json
