### 2023-10-23 53950b6
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/psf/black: 23.9.1 → 23.10.0](https://github.com/psf/black/compare/23.9.1...23.10.0)
- [github.com/pre-commit/mirrors-clang-format: v16.0.6 → v17.0.3](https://github.com/pre-commit/mirrors-clang-format/compare/v16.0.6...v17.0.3)
