### 2023-11-03 #5735 @shahmoradi
Thank you for your efforts to create this library. I am sure the Python community will find this package helpful. Everything about this project and JOSS submission, except the following comments I have for the authors:
1. The **installation** page should likely go before the **usage** page in the documentation website.
2. Adding a complete MPI-parallel example in the documentation would be helpful.
3. Instructions for running the unit tests (serial/parallel) would be very helpful (I am sure many, like me, do not know or remember how to run tests off the top of their heads).
4. A similar argument, but less severely, applies to docs generation. The instructions are particularly important and vital for future contributors.
5. Can the authors check if "Community guidelines: 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" exist in the repository and if so, point me to those docs?
6. Are there any alternative options available in the literature that the authors could cite in "State of the field: Do the authors describe how this software compares to other commonly-used packages?"
