+++ README.md
@@ -59,6 +59,44 @@ This is a minimal usage example of the `parafields` package:
 
 For more examples, check out the [parafields documentation](https://parafields.readthedocs.io/).
 
+## Reporting Issues
+
+If you need support with `parafields` or found a bug, consider a bug on
+[the issue tracker](https://github.com/parafields/parafields/issues).
+
+## Contributing
+
+`parafields` welcomes external contributions. For the best possible contributor
+experience, consider opening an issue on [the issue tracker](https://github.com/parafields/parafields/issues)
+before you start developing. Announcing your intended development in this way allows us to clarify
+whether it is in the scope of the package. Contributions are then done via a pull
+request on the GitHub repository. Please also add your name to the list of copyright holders.
+
+For a development installation of `parafields`, use the following instructions:
+
+```bash
+git clone https://github.com/parafields/parafields.git
+cd parafields
+python -m pip install -v --editable .[tests,docs,jupyter]
+```
+
+Before contributing, make sure that the unit tests pass and that new functionality is
+covered by unit tests. The unit tests can be run using pytest:
+
+```bash
+# Sequential tests
+python -m pytest
+
+# Parallel tests
+mpirun --oversubscribe -np 4 python -m pytest --only-mpi
+```
+
+In order to locally build the Sphinx documentation, use the following commands:
+
+```bash
+sphinx-build -t html ./doc ./html
+```
+
 ## Acknowledgments
 
 The [parafields-core C++ library](https://github.com/parafields/parafields-core) is
