### 2023-11-09 263ce4b
Add parapraph about other packages doing similar stuff
