+++ paper.bib
@@ -120,8 +120,6 @@ abstract = {This paper presents the basic concepts and the module structure of t
   doi          = {10.5281/zenodo.2565368},
   url          = {https://doi.org/10.5281/zenodo.2565368}
 }
-
-
 @article{wood1994simulation,
   title={Simulation of stationary Gaussian processes in [0, 1] d},
   author={Wood, Andrew TA and Chan, Grace},
@@ -132,4 +130,16 @@ abstract = {This paper presents the basic concepts and the module structure of t
   year={1994},
   publisher={Taylor \& Francis},
   doi={10.2307/1390903}
+}
+@article{davies2013lgcp,
+ title={On Circulant Embedding for Gaussian Random Fields in R},
+ volume={55},
+ url={https://www.jstatsoft.org/index.php/jss/article/view/v055i09},
+ doi={10.18637/jss.v055.i09},
+ abstract={The high-dimensionality typically associated with discretized approximations to Gaussian random fields is a considerable hinderance to computationally efficient methods for their simulation. Many direct approaches require spectral decompositions of the associated covariance matrix and so are unable to complete the solving process in a timely fashion, if at all. However under certain conditions, we may construct block-circulant versions of the covariance matrix at hand thereby allowing access to fast-Fourier methods to perform the required operations with impressive speed. We demonstrate how circulant embedding and subsequent simulation can be performed directly in the R language. The approach is currently implemented in C for the R package &amp;lt;b&amp;gt;RandomFields&amp;lt;/b&amp;gt;, and used in the recently released package &amp;lt;b&amp;gt;lgcp&amp;lt;/b&amp;gt;. Motivated by applications dealing with spatial point processes we restrict attention to stationary Gaussian fields on R&amp;lt;sup&amp;gt;2&amp;lt;/sup&amp;gt;, where sparsity of the covariance matrix cannot necessarily be assumed.},
+ number={9},
+ journal={Journal of Statistical Software},
+ author={Davies, Tilman M. and Bryant, David},
+ year={2013},
+ pages={1–21}
 }
