### 2023-11-09 fe589db
Put MPI documentation into Sphinx instead of notebook
