### 2023-11-09 #5735 @dokempf
@shahmoradi Thanks a lot for your comments, much appreciated.

>     1. The **installation** page should likely go before the **usage** page in the documentation website.

Done in `main`.

>     2. Adding a complete MPI-parallel example in the documentation would be helpful.

I have moved the MPI documentation from a Jupyter notebook (that cannot really be run in parallel) to the Sphinx documentation. Added additional bit especially about the differences between sequential and parallel use (same API, different shapes of the results of `evaluate`)

>     3. Instructions for running the unit tests (serial/parallel) would be very helpful (I am sure many, like me, do not know or remember how to run tests off the top of their heads).

Done in `main`.

>     4. A similar argument, but less severely, applies to docs generation. The instructions are particularly important and vital for future contributors.

Done in `main`.

>     5. Can the authors check if "Community guidelines: 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" exist in the repository and if so, point me to those docs?

Done in `main`. They are in the README which is also used to render the introduction section of the Readthedocs page.

>     6. Are there any alternative options available in the literature that the authors could cite in "State of the field: Do the authors describe how this software compares to other commonly-used packages?"

We added a paragraph about related (software) work to the paper draft.
