+++ paper.bib
@@ -143,3 +143,15 @@ abstract = {This paper presents the basic concepts and the module structure of t
  year={2013},
  pages={1–21}
 }
+@article{robbe2023grfjl, doi = {10.21105/joss.05595}, url = {https://doi.org/10.21105/joss.05595}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {89}, pages = {5595}, author = {Pieterjan Robbe}, title = {GaussianRandomFields.jl: A Julia package to generate and sample from Gaussian random fields}, journal = {Journal of Open Source Software} }
+@Article{mueller2022gstools,
+AUTHOR = {M\"uller, S. and Sch\"uler, L. and Zech, A. and He{\ss}e, F.},
+TITLE = {\texttt{GSTools} v1.3: a toolbox for geostatistical modelling in Python},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {15},
+YEAR = {2022},
+NUMBER = {7},
+PAGES = {3161--3182},
+URL = {https://gmd.copernicus.org/articles/15/3161/2022/},
+DOI = {10.5194/gmd-15-3161-2022}
+}
