### 2023-11-14 #5735 @gchure
> Hi @gchure do you have time to review this paper?

Hi  @diehlpk, I can review. This will be my first review with JOSS, so I would appreciate it if you could point me in the direction of some best practices.
