### 2023-11-16 #5735 @diehlpk
@gchure please have a look at the guidelines here

https://joss.readthedocs.io/en/latest/review_criteria.html

I am happy to answer all questions.
