### 2023-11-18 parafields/parafields#141 @gchure
JOSS Review -- Paper Comments
Hi @dokempf, I'm [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/5735) and am opening issues as I go along through my checklist. These comments pertain to the paper itself.

Overall, I think the paper meets all requirements for JOSS and I have no blocking change suggestions. Take the following as suggestions from someone out-of-field  who would appreciate a bit more context.

## Suggested Additions.
1. Lines `5-6`:

> "Application areas include e.g. soil physics, computer graphics and
> material sciences."

 I would appreciate more concrete examples than entire domains of science. Could you reference a few papers where the generation of stationary Gaussian fields are important in development of physical models?

2. Lines `28-30`:

> "In comparison with these alternative packages, parafields provides a rich feature set and the unique possibility to create Gaussian processes in a an MPI-distributed fashion."

 I would appreciate specific examples of what is provided by parafields that `GSTools` or `lgcp` do not have, beyond being able to employ MPI distribution.

3. Line `55`:

> "In order to embed perfectly into existing user workflows[...]"

Perfection is in the eye of the beholder, so I would tone down this sentence. You furthermore mention (lines `40-41`) that you put "special emphasis on [...] embedding into existing user workflows." Having these output as numpy-arrays is not sufficient for me to conclude you put special emphasis on this aspect. I would suggest removing this phrasing or dropping discussion of workflow integration entirely.

## Typos
I found a few typos that should be corrected.
1. Line `14`: "in particular if the" -> "particularly if"
2. Line `44`. "systen" -> "system"
3. Mixed usage of British and American English (e.g. line `19` "factorisation" and line `41` "customizability")

That's it! Thanks for the clear and concise work.
