### 2023-11-18 #5735 @gchure
I'm opening issues on the parafields repository as I go along through my review. Here are comments on the paper text: https://github.com/parafields/parafields/issues/141

Note that my paper comments are non-blocking and I leave it up to the discretion of @dokempf.
