### 2023-11-18 parafields/parafields#142 @gchure
JOSS Review -- Non-reproducible example implementing parallel usage
Hi @dokempf, I'm [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/5735) and am opening issues as I go along through my checklist. These comments pertain to replication of the examples on the documentation.

I've been able to replicate all examples on your documentation locally, save for your [verbatim example for basic parallel usage](https://parafields.readthedocs.io/en/latest/parallel.html#basic-usage).

Running the following script `test.py`:

```
from mpi4py import MPI
import parafields
field = parafields.generate_field(
    cells=(256, 256), extensions=(1.0, 1.0), covariance="exponential", variance=1.0
)
print(field.shape)
```

as

```
> mpiexec -n 4 python test.py
```

generates the following traceback:

```
Traceback (most recent call last):
  File "/Users/gchure/Desktop/test.py", line 6, in <module>
Traceback (most recent call last):
  File "/Users/gchure/Desktop/test.py", line 6, in <module>
Traceback (most recent call last):
  File "/Users/gchure/Desktop/test.py", line 6, in <module>
    print(field.shape)
Traceback (most recent call last):
  File "/Users/gchure/Desktop/test.py", line 6, in <module>
    print(field.shape)
AttributeError: 'RandomField' object has no attribute 'shape'
    print(field.shape)
AttributeError: 'RandomField' object has no attribute 'shape'
AttributeError: 'RandomField' object has no attribute 'shape'
    print(field.shape)
AttributeError: 'RandomField' object has no attribute 'shape'
```

This is running `Python==3.10.9`, `parafields==0.3.0`, `mpi4py==3.1.5` using `MPICH==4.1`.
