### 2023-11-18 #5735 @gchure
Here is an open issue [(#142)](https://github.com/parafields/parafields/issues/142) outlining a error in reproducing the example for parallel usage of `parafields`. This is a blocking issue that should resolved before formal publication.
