### 2023-11-18 parafields/parafields#143 @gchure
JOSS Review -- Remove option for 3-dimensional GRF in interactive mode
Hi @dokempf, I'm [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/5735) and am opening issues as I go along through my checklist. This comment pertains to a feature suggestion for interactive mode.

I found your tool for interactively generating random fields to be i) fun and ii) very useful. However, you provide an option for 3-dimensional field specification that has no actual utility in interactive mode:

Why not just remove Dimension: 3 as an option in the dropdown?  I think [this is the offending line](https://github.com/parafields/parafields/blob/84c17d4399a7b3f24df90e87c9e397418b692bf4/src/parafields/interactive.py#L160C1-L160C58) that could just use an extra logic test to prevent its population.
