### 2023-11-18 #5735 @gchure
Here is an open issue [(#143)](https://github.com/parafields/parafields/issues/143) which highlights an unsupported option in interactive mode. This is non-blocking and is at  @dokempf's discretion.
