### 2023-11-18 parafields/parafields#144 @gchure
JOSS Review -- Improve test coverage for FEniCSx integration
Hi @dokempf, I'm [reviewing your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/5735) and am opening issues as I go along through my checklist. These comments pertain to test coverage.

The parafields documentation highlights [integration with FEniCSx-based PDE solvers](https://parafields.readthedocs.io/en/latest/intro.html) and [even gives an example of how to use it](https://github.com/parafields/parafields/blob/main/jupyter/fenicsx.ipynb), yet the relevant function:
https://github.com/parafields/parafields/blob/84c17d4399a7b3f24df90e87c9e397418b692bf4/src/parafields/field.py#L736

[has zero test coverage](https://app.codecov.io/gh/parafields/parafields/blob/main/src%2Fparafields%2Ffield.py#L736). As this is advertised as a feature, it should have some code coverage.
