### 2023-11-18 #5735 @gchure
Here is an open issue [(#144)](https://github.com/parafields/parafields/issues/144) which highlights non-tested code that is advertised as a feature of `parafields`. I personally think that judging testing over coverage percent is silly (which is why I'm fine with 80% coverage), but the fenicsx integration function seems to be entirely untested. I don't think this should be a blocking issue, but really should be addressed. I leave that up to @diehlpk's discretion.

This is my final review comment. Once these issues are closed, I'm happy to recommend it be formally published in JOSS.
