### 2023-11-27 d89e675
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/psf/black: 23.10.1 → 23.11.0](https://github.com/psf/black/compare/23.10.1...23.11.0)
- [github.com/python-jsonschema/check-jsonschema: 0.27.1 → 0.27.2](https://github.com/python-jsonschema/check-jsonschema/compare/0.27.1...0.27.2)
- [github.com/pre-commit/mirrors-clang-format: v17.0.4 → v17.0.5](https://github.com/pre-commit/mirrors-clang-format/compare/v17.0.4...v17.0.5)
