+++ .github/workflows/joss.yml
@@ -21,7 +21,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v3
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
