### 2023-12-01 d745de9
Switch to PyPI trusted publishing
