### 2023-12-08 parafields/parafields#143 @dokempf
The rationale here was that while a visualization is not available, the widget UI still gives an intuitive grasp of the available parameter space. I would therefore like to keep it.
