+++ paper.bib
@@ -155,3 +155,76 @@ PAGES = {3161--3182},
 URL = {https://gmd.copernicus.org/articles/15/3161/2022/},
 DOI = {10.5194/gmd-15-3161-2022}
 }
+
+
+
+@article{cameletti2013spatio,
+  title={Spatio-temporal modeling of particulate matter concentration through the SPDE approach},
+  author={Cameletti, Michela and Lindgren, Finn and Simpson, Daniel and Rue, H{\aa}vard},
+  journal={AStA Advances in Statistical Analysis},
+  volume={97},
+  pages={109--131},
+  year={2013},
+  publisher={Springer}
+}
+
+
+@article{sain2011spatial,
+  title={A spatial analysis of multivariate output from regional climate models},
+  author={Sain, Stephan R and Furrer, Reinhard and Cressie, Noel},
+  journal={The Annals of Applied Statistics},
+  pages={150--175},
+  year={2011},
+  publisher={JSTOR}
+}
+
+
+
+@article{dodwell2015hierarchical,
+  title={A hierarchical multilevel Markov chain Monte Carlo algorithm with applications to uncertainty quantification in subsurface flow},
+  author={Dodwell, Tim J and Ketelsen, Christian and Scheichl, Robert and Teckentrup, Aretha L},
+  journal={SIAM/ASA Journal on Uncertainty Quantification},
+  volume={3},
+  number={1},
+  pages={1075--1108},
+  year={2015},
+  publisher={SIAM}
+}
+
+
+@article{penny2005bayesian,
+  title={Bayesian fMRI time series analysis with spatial priors},
+  author={Penny, William D and Trujillo-Barreto, Nelson J and Friston, Karl J},
+  journal={NeuroImage},
+  volume={24},
+  number={2},
+  pages={350--362},
+  year={2005},
+  publisher={Elsevier}
+}
+
+
+
+@article{torquato2002random,
+  title={Random heterogeneous materials: microstructure and macroscopic properties},
+  author={Torquato, Salvatore and Haslach Jr, HW},
+  journal={Appl. Mech. Rev.},
+  volume={55},
+  number={4},
+  pages={B62--B63},
+  year={2002}
+}
+
+
+
+@article{scheichl2017quasi,
+  title={Quasi-Monte Carlo and multilevel Monte Carlo methods for computing posterior expectations in elliptic inverse problems},
+  author={Scheichl, Robert and Stuart, Andrew M and Teckentrup, Aretha L},
+  journal={SIAM/ASA Journal on Uncertainty Quantification},
+  volume={5},
+  number={1},
+  pages={493--518},
+  year={2017},
+  publisher={SIAM}
+}
+
