+++ paper.bib
@@ -227,4 +227,3 @@ DOI = {10.5194/gmd-15-3161-2022}
   year={2017},
   publisher={SIAM}
 }
-
