### 2023-12-08 parafields/parafields#143 @gchure
That sounds fair. I think it would help if you acknowledge this with an `attention` admonition on the [documentation page](https://parafields.readthedocs.io/en/latest/interactive.html).
