### 2023-12-11 parafields/parafields#141 @dokempf
Incorporated into the `joss-paper` branch.
