### 2023-12-11 parafields/parafields#144 @dokempf
For context: The reason this is currently not properly tested in CI is that FEniCS is not `pip`-installable and it would require a change (to the worse) to the dev installation instructions. You are of course right that this should be done @gchure, but I went with Patrick's suggestion to document the testing status for now (and hope for `pip`-installable FeniCSx)
