+++ pyproject.toml
@@ -18,7 +18,7 @@ build-backend = "setuptools.build_meta"
 # single source of truth for parafieldsseq's version
 [project]
 name = "parafields"
-version = "0.3.0"
+version = "1.0.0"
 description = "Parallel Parameter Fields for Uncertainty Quantification"
 readme = "README.md"
 maintainers = [
