### 2023-12-11 cbf8ce4
Remove setuptools pin
