### 2023-12-11 parafields/parafields#144 @gchure
This sounds good to me, and I'll keep my fingers crossed for `pip install fenicsx`.
