### 2023-12-15 #5735 @diehlpk
The title on Zenodo seems to be wrong

parafields/parafields: parafields v1.0.1

There should be the same title as on the paper.
