### 2023-12-15 #5735 @diehlpk
> > @diehlpk  All done, The DOI is [10.5281/zenodo.10391370](https://doi.org/10.5281/zenodo.10391370)
>
> And the version of the release?

Nevermimd, I found the version in the wrong Zenodo title.
