### 2023-12-15 #5735 @diehlpk
Oh, the title should be **parafields: A generator for distributed, stationary Gaussian processes** or?

Please add that to the paper and archive.
