### 2023-12-18 #5735 @dokempf
@diehlpk Finished, here is the DOI: [10.5281/zenodo.10355636](https://zenodo.org/doi/10.5281/zenodo.10355636)
