+++ paper.bib
@@ -165,7 +165,8 @@ DOI = {10.5194/gmd-15-3161-2022}
   volume={97},
   pages={109--131},
   year={2013},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10182-012-0193-3}
 }
 
 
@@ -175,7 +176,8 @@ DOI = {10.5194/gmd-15-3161-2022}
   journal={The Annals of Applied Statistics},
   pages={150--175},
   year={2011},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.1214/10-AOAS369}
 }
 
 
@@ -188,7 +190,8 @@ DOI = {10.5194/gmd-15-3161-2022}
   number={1},
   pages={1075--1108},
   year={2015},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/130915005}
 }
 
 
@@ -200,7 +203,8 @@ DOI = {10.5194/gmd-15-3161-2022}
   number={2},
   pages={350--362},
   year={2005},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2004.08.034}
 }
 
 
@@ -212,7 +216,8 @@ DOI = {10.5194/gmd-15-3161-2022}
   volume={55},
   number={4},
   pages={B62--B63},
-  year={2002}
+  year={2002},
+  doi={10.1115/1.1483342}
 }
 
 
@@ -225,5 +230,6 @@ DOI = {10.5194/gmd-15-3161-2022}
   number={1},
   pages={493--518},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/16m1061692}
 }
