+++ paper.bib
@@ -166,7 +166,7 @@ DOI = {10.5194/gmd-15-3161-2022}
   pages={109--131},
   year={2013},
   publisher={Springer},
-  doi={10.1007/s10182-012-0193-3}
+  doi={10.1007/s10182-012-0196-3}
 }
 
 
