### 2023-12-18 #5735 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
