+++ paper.bib
@@ -12,7 +12,7 @@ keywords = {Inverse modeling, Geostatistical inversion, Uncertainty quantificati
 abstract = {In the geostatistical inverse problem of subsurface hydrology, continuous hydraulic parameter fields, in most cases hydraulic conductivity, are estimated from measurements of dependent variables, such as hydraulic heads, under the assumption that the parameter fields are autocorrelated random space functions. Upon discretization, the continuous fields become large parameter vectors with O(104−107) elements. While cokriging-like inversion methods have been shown to be efficient for highly resolved parameter fields when the number of measurements is small, they require the calculation of the sensitivity of each measurement with respect to all parameters, which may become prohibitive with large sets of measured data such as those arising from transient groundwater flow. We present a Preconditioned Conjugate Gradient method for the geostatistical inverse problem, in which a single adjoint equation needs to be solved to obtain the gradient of the objective function. Using the autocovariance matrix of the parameters as preconditioning matrix, expensive multiplications with its inverse can be avoided, and the number of iterations is significantly reduced. We use a randomized spectral decomposition of the posterior covariance matrix of the parameters to perform a linearized uncertainty quantification of the parameter estimate. The feasibility of the method is tested by virtual examples of head observations in steady-state and transient groundwater flow. These synthetic tests demonstrate that transient data can reduce both parameter uncertainty and time spent conducting experiments, while the presented methods are able to handle the resulting large number of measurements.}
 }
 @article{dietrich1997fast,
-  title={Fast and exact simulation of stationary Gaussian processes through circulant embedding of the covariance matrix},
+  title={Fast and exact simulation of stationary {G}aussian processes through circulant embedding of the covariance matrix},
   author={Dietrich, Claude R and Newsam, Garry N},
   journal={SIAM Journal on Scientific Computing},
   volume={18},
@@ -34,6 +34,7 @@ abstract = {In the geostatistical inverse problem of subsurface hydrology, conti
   doi={10.1137/17m1149730}
 }
 @misc{parafields-core,
+  title={{parafields-core}},
   author={Klein, Ole and Kempf, Dominic},
   year={2022},
   publisher={GitHub},
@@ -89,13 +90,13 @@ abstract = {This paper presents the basic concepts and the module structure of t
    author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
    year = {2017},
    note = {https://github.com/pybind/pybind11},
-   title = {pybind11 -- Seamless operability between C++11 and Python}
+   title = {pybind11 -- Seamless operability between {C}++11 and {P}ython}
 }
 @misc{dune-randomfield,
    author = {Ole Klein},
    year = {2017},
    note = {https://gitlab.dune-project.org/oklein/dune-randomfield},
-   title = {dune-randomfield - Generation of Gaussian random fields in arbitrary dimensions, based on circulant embedding}
+   title = {dune-randomfield - Generation of {G}aussian random fields in arbitrary dimensions, based on circulant embedding}
 }
 @misc{fakempi,
    author = {Dominic Kempf and the PetSc Developers},
@@ -110,9 +111,9 @@ abstract = {This paper presents the basic concepts and the module structure of t
                   Max Smolens and
                   Michael Grauer and
[…]
