+++ paper.bib
@@ -154,7 +154,7 @@ publisher = {The Open Journal},
 volume = {8},
 number = {89},
 pages = {5595},
-author = {Pieterjan Robbe}, 
+author = {Pieterjan Robbe},
 title = {GaussianRandomFields.jl: A {J}ulia package to generate and sample from {G}aussian random fields},
 journal = {Journal of Open Source Software}
 }
