### 2023-12-18 #5735 @danielskatz
@dokempf - I've suggested some small changes in https://github.com/parafields/parafields/pull/153. Please merge this, or let me know what you disagree with.

In addition, please add countries to the affiliations in the paper. I'm also confused by the casing for parafields, which is "parafields" in the title, but "Parafields" in the paper in a number of places, such as the start of the first and second paragraphs.
