### 2023-12-18 #5735 @dokempf
@danielskatz Thanks a lot for your suggestions, which I have merged into the paper. Also added the requested geographic information. Regarding capitalization: parafields is only capitalized at the start of the sentence, where I though it is mandatory, but please correct me if I am wrong. As a german, I typically struggle with capitalization in english because we have quite unique capitalization rules...
