### 2023-12-18 #5735 @danielskatz
There is no consistent rule for this. There are many style guides, and they are not consistent. One option suggested is to try to move the term away from the start of the sentence. In this case, I'm just going to go ahead and accept it the way it is.
