### 2023-12-18 #5735 @danielskatz
Congratulations to @dokempf (Dominic Kempf) and co-authors on your publication!!

JOSS depends on volunteers, and we couldn't do this without you.
